Show spinner and disable buttons on form submit in notifications list
There was no feedback to the user when marking notifications
as read and unread. When marking a bigger amount of notifications
it can take a bit to update the view, which might lead to addtional
submit attempts. By showing a spinner and disabling the button
we can prevent this from happening.

Co-authored-by: Eduardo Navarro <enavarro@suse.com>