Prevent exception on empty URL in markdown text
An empty URL part in markdown text leads to an exception
in the mardown renderer. When a comment
includes such an empty URL string, it ends up with a
code 500. This happens when people post code examples
which contain the same syntax as a markdown link.
Instead of throwing an exception, we can just render
the original content incase of an empty URL part.

Kudos to @eduardoj

Fixes #10259