Prevent rendering of carriage return character in xml views
When editing the description of a package/project in the webui,
the text_area in the form add's a sequence of '\r\n' to indicate
a new line. The carriage return indicated by the '\r' gets
rendered in the corresponding xml output for example in the meta
view or through the API. In order to prevent this from happening
we can replace this sequence with a simple '\n' which is the
standard in most recent unix environments to indicate a new line.

Fixes #10272