Wrap Left Navigation items text into span tag
The commit wraps text for items in the Left Nav menu into span tag and
adds an appropriate class for the items, so that it allows to find the
text using CSS selectors and thus allows to manipulate with the text.

Also, the commit adds title attribute for the links in Left Navigation
menu. That is needed for the case when sidebar is collapsed to show
tooltip while hovering over the items.