Prevent duplication for involved projects/packages
If the Contribution Graph is enabled, the 'Involved Projects/Packages'
list can be displayed inside a simple card or inside a tab according to
the viewport size.

Instead of rendering the same partial ('_involvement.html.haml') twice
and using CSS to show/hide one or the other according to the viewport
size, now we inject the content of the partial in the right place using
JavaScript.

Preventing the duplication is needed because it brings conflicts when
selecting other elements in the view by JavaScript.

Co-authored-by: Eduardo Navarro <enavarro@suse.com>