Wrap sidebar links with span.nav-item-name
The PR #10263 moved links for account managing to sidebar. Since sidebar
became collapsible in #10293, those links have to be wrapped into
span.nav-item-name. Also, title should be specified for them.

The commit adds those missing wrappers and titles, and also capitalizes
the links to make them consistent with the other links in sidebar.