[backend] speed up removal of old projpacks_linked data
We now maintain a new projpacks_linked_blks hash that stores
the size of the same-project blocks in the corresponding
projpacks_linked entry. Using this we can for one skip the
blocks much faster if the project does not match and we
also can remove the block with one single splice call.

This speeds up the removal phase considerably. It no
longer dominates the time used in setup_projects.