Disable comment submit button if the body is empty
On an empty comment being in the "Write" tab and then clicking "Update"
or "Add" resulted in "Please fill out this field" warning. But there
were no warning when the comment was empty and "Preview" tab was
selected.

The commit addresses the issue by disabling submit button until text
area is filled in.