[backend] allow calling the sign tool with a bad signkey
The code used to be different in the publisher and the signer.
The publisher used to ignore bad signkeys (i.e. too small
signkeys), whereas the signer just used the keys. This was
changed in commit 27cfcb1aaffbf0e0ba2c0cf06ce87a0373c0a563
which made both versions of the code ignore bad signkeys.

It turned out that doing this has the disadvantage that we
do not catch configuration errors anymore which used to
result in the signer complaining about a bad signkey.
So this coomit changes the code in both the signer and the
publisher to again pass bad signkeys to the signer.