Add characters count for User Biography
There is a limit of characters that can be typed in Biography textarea.
After the limit is reached, no characters can be entered. But there is
no visual hint on how many characters remaining.

The commit adds a counter that updates in real-time to show to the user
on how many characters remaining.