Fix showing elements of an image
Fixes an strange behaviour of showing different description elements.
This happened not when the Details section was accessed directly with an URL,
but when it was accessed clicking in the Details tab after clicking the
Software tab.

Show always the description (specification) of an image.
Show Author and Contact fields only in Details tab.