Fix worker status data inserted in status_history table
A previous code refactoring introduced a change in the amount of registers
inserted to be used in the system status graphic. This change made the graphic
not to update correctly. See 1bde603dd744325de326e501bc27151ca65d5f37.

Revert part of this refactoring, to ensure that '0' values are inserted.
This way the `resample` helper will work again with datasets with the same time
limits.