Remove mentions of openFATE on error pages
openFATE has been discontinued and should no longer be linked to as
there is no archive.

This commit is replacing the outdated references with the same text as
used in https://github.com/openSUSE/obs-landing/pull/276 referencing
github issues as the correct place for features as well as used for
issue reports.

Fixes #10566

Co-authored-by: Oliver Kurz <okurz@suse.de>