Adjust the rollout tasks
Most of the rollout tasks like "all_on" or "from_beta" excluded the staff
users. Semantically speaking it was confusing. As we no longer need that
restriction, the staff are not excluded anymore.

We also add tasks to move the anonymous user (_nobody_) into or out of the
rollout program. Something that is usually needed at the beginning or end of
the rollout process.