Remove unneeded path parameter for creating an ...
... issue tracker from apidocs documentation.

The `POST` route for issue trackers doesn't require an issue tracker id
in path parameters. And the `create` method in the controller doesn't
make use of any path parameters.

Also replace `bnc` as an example of ids with `<id>`, in the comments,
for the description of the route related to the controller method.