[backend] scheduler: move linked project check after getprojpack
We tested for linked projects at the start of do_fetchprojpacks,
which created a race if the link information was updated while
the getprojpack rpc was running.

We now do the test after the getprojpack call returns.

We also add the linked packages we know about right away to
reduct the number of rpcs needed.