Generate download_name from binarylist
This is the first step to fix #10572 by generating the download_name of
a package from available binaries.

We iterate over the binaries currently available for the current package
and skip everything that is not an .rpm file.

Example: We skip "rpmlint.log", "_buildenv" etc. but have a look at
"name-version-arch.rpm".

We remove the file extension and everything including the second dash
from the end of the filename.

Example: "hello-2.10-2.3.src.rpm" is shortened to "hello".

We add each entry to the list of download_name_candidates. Once we are
done iterating over the binaries, we select the shortest entry.

This is necessary because some packages actually have a list of
binaries whose names don't correspond with the actual package
people should download from software.opensuse.org.

Example: A package "MyPackage" can have multiple binaries with a
variety of names such as "my-package", "my-package-debug",
"my-package-devel" etc. but we are only interested in "my-package".