Show correct message when hovering build info elements
After introducing another element with CSS class 'toggle-build-info' in
a previous commit, clicking in one of this elements should change the
title for all elements which have this class.

Prevent having one build info element with title 'Click to keep it open'
and another element with title 'Click to close it' at the same time.