[backend] do dodfetch/dodcheck even if download_on_demand is not enabled
A download on demand repo can be on another partition or even
on a remote instance. So we need to always do the dodcheck so
that all needed dod binaries are available if a build job is
created.
Otherwise the workers do the dod fetching which leads to the
getbinaryversion calls blocking everything else.