Enable responsive_ux feature flag before each RSpec spec
It has to be before each spec since the Flipper feature flags are
cleaned by database_cleaner between specs. We do sometimes enable a
feature flag only for a few specs and this has to be cleaned up to not
affect other specs. Preventing database_cleaner from cleaning up the
Flipper feature flags is therefore not a valid option.