[backend] remove container on service timeout
Without this patch on a service timeout call-service-in-container
process gets killed but container is still running. Some services
use locking and will block other service forever if the container
does not get removed properly

This patch ensures that the container gets stopped and removed
if call-service-in-container process gets killed by bs_service.