Remove remaining show JS view file
This file should have been removed when the inline editing was modified
and the cancel button no longer redirected to show.js.erb.
Moreover, the show.js.erb called a partial but one of the locals was
missing.

Keeping this file causes errors when bots try to visit the former URL.

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>