Refactor code for tokens
Result of multiple mob programming sessions to lay out foundations to a
better SCM/CI integration.

The controllers trigger_controller and services/webhooks_controller are
now merged and a single action is handling the various tokens. The
routes are kept for backward compatibility.

Authorization is done through Pundit, although we couldn't get away from
Package.get_by_project_and_name... it's just doing too much and we
wanted to avoid increasing the scope of the changes even more.

Anything related to extracting tokens out of the various
headers/parameters is now handled by
TriggerControllerService::TokenExtractor. It was previously scattered in
the aforementioned controllers and token models.

Finally, the specs were adapted to follow the changes.

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>
Co-authored-by: Dany Marcoux <dmarcoux@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>
Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>
Co-authored-by: Lukas Krause <lkrause@suse.de>
Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>
Co-authored-by: Victor Pereira <vpereira@suse.com>