Extract data from a incoming SCM's webhook
This kind of webhooks trigger the new Token::Workflow.
They come from SCMs like GitHub or GitLab.

We extract the data from the payload of the incoming webhook/event,
then check that the event and data are valid, and add it to the token
as attributes.

We create a new controller with the behaviour of the token workflow.

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>