[backend] implement cross build support in the scheduler
If we're doing a cross build we set up a second package pool
for the host repositories. We also split the build requires
into host/sysroot dependencies and use get_sysroot() for
expansion.

Packages going into the sysroot are marked with "sysroot" in
the build job. The host repositories are put into the
"syspath" elements, like with kiwi image builds.

Note that you always need both the native architecture and
the target architecture in the repository. The native
architecture is used for native package builds and also
to forward host repository changes to the target architecture
scheduler.