Skip validate_params for both GitHub and GitLab
After more detailed tests, it is clear that the `validate_params` can
fail in both GitHub and GitLab webhooks. Each of them can send a
non-string value as parameters.

For example, GitHub sends an integer value for "number" key (PR number)
and GitLab sends a hash representing the project, the user, etc.