Execute the workflow's steps
After downloading the workflow's configuration file we execute the
workflow's steps against the configured project and package.

Then, validate the steps inside the Step model.

Depending on each kind of step, we know which kind of event and actions
we expect from the webhooks and what actions we want to perform or
ignore (i.e. branch a packae in OBS, just rebuild the package, etc.)

That's why we moved the checks about webhook events and actions to the
Step class.

Co-authored-by: Eduardo Navarro <enavarro@suse.com>