Rename "linked projects" to "project with a link"
... in the TriggerController tests.

"project" and "package" refer to a local project and a local package.
"project_with_a_link" refers to a project that could link to a local
project, or a remote project.
"local_package" is a local package, created in a project with a link.

Regenerate cassettes for the tests changed.