Fix the style and the collapse of Mentioned Issues
Collapsing the summary of the issue was broken. It was implemented by
Webui::WebuiHelper#toggle_sliced_text. The helper has been removed and
the text is now collapsed in the same way as any other text in the application,
with "Show more/less" link.

Improve the style of the section too.