Revert "Apply the background color all the way ...
... to the right" and "Only create a CSS grid on code inside diffs".

This reverts commit 55d02ddc767507efc2d6ec0aa134c290341c8206 and commit
7d564eb9130cbd6d6ed7e29fb0a0a6e2e759d122.

Creating a work around with CSS grid to apply a background color all the
way to the right breaks diffs with more than 1000 lines, both in Chrome
and Chromium. And breaks diffs with more than 10000 lines in Firefox.

This solution could be used for small diffs, but not for all of them.