[backend] support DoD container registries
Those DoD resources are different to repositories because there
is no way to request the complete state from the server.

Thus we collect the needed DoD resources in the scheduler and put
them into a "doddata.needed" file. The bs_dodup server then uses
this information to request data for each resource.