Report back build state to the correct scm repository
We were using the head (source) repository instead of
the base (target) repository to report back the
build state of a workflow. This is wrong since the
state is reported using the scm token of the target
repo owner/maintainer, which leads to permission
problems (since it doesn't associate the reported
state to the PR, but directly to the commit on the
source repository).

We also changed the download url, we should only
take the "workflow.yml" file from the target
repository.

Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>