Add updating actions to BranchPackageStep
When a PR is updated in the SCM, the rebuild of the package in OBS is
triggered automatically due to the _service functionality.  There is no
need to branch the package again.  On the workflow side, we only have to
update the subscription to report back to the right commit sha.

The methods that return a name are renamed to end with _name. For
example, target_project becomes target_project_name. A new method
target_package is created to return the object, not the name.