Refactor Token::Workflow model and add exception
Right now we have only one workflow and one or zero steps in it. But in
the near future will need to loop over many of them.
Starting looping already, will save us to check if the step exists. It
usually happens when the incoming event is one that we don't want to
handle. So, in this case, we just ignore it and do nothing.

From now on, an exception is raised when OBS isn't able to branch the
package for any reason.