Take user from the token, not from the session
If we use the `token.user.run_as` method from the `User` model before
calling `token.call`, then the user from the session is the same as the
user from the token. However if someone forgets to do so, then the user
from the session could be different. This would cause problems.

The changes are also fixing an error in the creation/update of event
subscriptions since we need to pass a user, not a user login.