Only set the acceptinfo oxsrcmd5/xsrcmd5 attr if the link can be expanded
The old code sets the oxsrcmd5/xsrcmd5 attribute of the <acceptinfo/>
element even if the corresponding link cannot be expanded. In this case,
the oxsrcmd5/xsrcmd5 refers to the unexpanded sources, which contradicts
the purpose of the oxsrcmd5/xsrcmd5.
In order to fix this, only set these attributes if the correponding
links can be expanded. If a link cannot be expanded,
BSSrcServer::Link::handlelinks either dies or returns an error
string, where the latter is not checked by the old code. Hence, only
set the oxsrcmd5/xsrcmd5 if handlelinks returns a reference.