[dist] optional dependencies for obs-api-support.target
Each service in listed in obs-api-support.target requires the database.
So it makes no sense to start before mariadb is up and obsapisetup
(which creates the database) finished.

The dependencies are optional because it depends on the setup if
database and api running on one machine. This is given in our appliance
but may differ in other setups.