Disable CodeClimate checks which have RuboCop equivalents
This is to avoid reporting twice the same thing due to having two
standards for various code quality metrics. RuboCop is the standard
whenever CodeClimate and RuboCop have the same checks/cops.

The CodeClimate checks are described here
https://docs.codeclimate.com/docs/maintainability#section-checks

I compared them to the cops provided by RuboCop to see which ones are
overlapping. This is how I came up with a list of CodeClimate checks to
disable.