Extract workflow validations into three custom validators
The event and action coming from the webhooks are no longer
validated inside the BranchPackageStep but at the Workflow level.

All the validations we have to do for Workflow are now splitted in
three custom validators:
- WorkflowEventAndActionValidator
- WorkflowStepsValidator
- WorkflowFiltersValidator

Rename Workflow#workflow method to Workflow#workflow_instructions.

Both Workflow and BranchPackageStep models are now ActiveModel and we
use attribute accessors. However, we need to overwrite their
initializers to include `with_indifferent_access` for the hashes.

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>
Co-authored-by: Dany Marcoux <dmarcoux@suse.com>
Co-authored-by: Rubhan Azeem <rubhan.azeem@suse.com>
Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>