Implement LinkPackageStep
The SCM integration allows to trigger a new step: LinkPackageStep.
With this, the user can create a package linked to another existing
source package. This functionality is based on `osc linkpac`.

We add a package called `_project` with a `_service` file inside. This
is a temporary solution to be able to trigger the services in a linked
package.

Adjust specs and cassettes accordingly.

Co-authored-by: Rubhan Azeem <rubhan.azeem@suse.com>