After upgrading Chromium, fix specs on mobile ...
... by scrolling to the link.

Upgrading Chromium fixed a false positive in feature package tests, when
adding a user or adding a group, making the tests fail. The bad behavior
was that clicking the links 'Add User' or 'Add Group' was possible, even
if the link was outside the visual area.

Scroll to those links in these failing tests in mobile. This makes
feature tests to pass again.

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>