[backend] Add `tos_url` to configuration
The OBS frontend introduced an new field to the configuration
table (https://github.com/openSUSE/open-build-service/pull/11437)
in order to save a URL for the terms of services.
The frontend writes this configuration to the backend through
a delayed job. The backend currently throws an 400 http code,
since it is not aware of the new field.

Fixes #11436