Delete project and package on destroy callback from sphinx indexes
Add the moment we only populate new records to sphinx. But we
don't remove those indexes after those records get deleted from
the database.
Usually the Thinking Sphinx gem is taking care of this automatically,
but since we differ from the regular setup by writing to the sphinx
index in a delayed job, we also have to consider the deletion.