Allow searching by real name in users DataTable
The existent "User" column contains a combination of the user's realname
and the user's login, but it is only possible to search by one of them,
"login" in this case. As it is not possible to search by both of them, a new
hidden column has been added to allow searching by user's realname too.