Discard StagingProjectAcceptJob on InvalidStateError
This can happen when at the same time of scheduling the job
there will be a new request selected to the staging project.
Or somehow a staged request changes state in the same time.

If we discard the job on this error, the staging project state
calcuclation will be done again.

Partially addresses #10636