Replace source_project by project in ConfigureRepositoriesStep
As a consequence, the 'source_project_name' is no longer required for
the three types of steps (BranchPackageStep, LinkPackageStep,
ConfigureRepositoriesStep). So the validation and definition of the
method are moved to the two first steps. On the third one, a new method
'project_name' and its validation is defined inside
ConfigureRepositoriesStep.