[dist] obsstoragesetup: do not prompt before wipe
According to lvcreate man, --wipesignature, "there is a prompt for
each signature detected to confirm its wiping (unless --yes is used to
override confirmations.)". However, obsstoragesetup is executed as a
service and it will abort if an interactive prompt is requested.

The result is that workers are started without a proper storage and
they will start to use rootfs, that might not be prepared to handle
the amount of worker data.

Use --yes for lvcreate to avoid interactve prompts.

Signed-off-by: Luiz Angelo Daros de Luca <luizluca@gmail.com>