Create web notification for subscribed group members
Having a single notification for a subscriber_type Group won't allow
users of this group to have their own notifications (like marking them
as read/unread).  We need to create a notification for every group
member.

By linking a notification to a group, we can tell which groups have
notifications. This is useful for UI filters.

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>
Co-authored-by: Lukas Krause <lkrause@suse.de>