Align the configuration of CodeClimate checks with those of RuboCop cops
Using the CodeClimate plugin for RuboCop is not possible due to our
somewhat quirky RuboCop setup with our 2 configurations in the Git
repository's root and `src/api/` directories.

This solution is good enough for now to allow us to track the code
quality changes over time with the CodeClimate UI.

We also exclude some more directories which we don't want to analyze.