Simplify comments' section header
Replace "Comments for X" with simply "Comments" instead of finding a way
to shorten the long names.

The comments are displayed in the context of a certain item (project,
package) so there is no need to repeat the item's name,
which is clear in other parts like the breadcrumbs or the title.

We also have comments in the request page but there it's useful to keep
the request's number they are related to.