Elide project/package names set in flash messages
The name of the project or package is usually on the page or was
previously added by the user into a form. So the user doesn't need to
rely on the flash message to know which project/package the application
is talking about.

So with elisions, we shorten the flash messages, which is especially
useful in small screens.