Cache the diff if no one ever tried
Before #11296 creating BsRequests via the UI never cached the diff.
Only looking at the BsRequest in the UI would do so. Since the PR
we always cache the diff on BsRequest creation.

But that means there can be requests that never have been looked at
but also never got BsRequestActionWebuiInfosJob scheduled for it.
So the cache will never be created, no matter how often you press the
refresh button.

In that case, let's schedule the jobs if there isn't a job in the queue
currently. Fixes #11770