Fix inconsistent target project name
The Workflow::Step::ConfigureRepositories#target_project_name method was
a redefinition of the Workflow::Step#target_project_name.

They should be identical except for the YAML key containing the project
name. So there is no need to keep both. In fact, this is leading to
inconsistencies. For example with the format of the name (using "-" or
":").

Now we only have one definition of the method. The only part we redefine
is the project name. For this, we have the `target_project_base_name`
method.