Reintroduce checking for 'Token' string
Token authentication can be performed by passing a token string through
an 'Authorization' HTTP header. The format to pass this token follows
this expression:

Token TOKEN_STRING

This changed in commit deea2dfcd9eea704eaed0532793160f7c7ba88b6, where
the check for the 'Token' string was removed.

This commit reintroduces the check for the 'Token' string, as was
initially defined, and as is described in the API documentation.