Remove duplicated Contribution Graph spec
We have the spec duplicated for both beta and non-beta version of the
code.

The non-beta one is causing flickering tests in our CI. Since nobody
sees the non-beta version of the UI anymore and the feature flag is
going to be removed in a few days, we decide to remove the non-beta spec.