[backend] add obsscm helpers to the BSSrcrep module
We store the extra data in the _serviceerror file if the
service is in progress. We put a "in progress" marker
both at the start and the end, so that the getserviceerror
function will pick up the correct state.

The data is stored in key=value lines where the values
are url-encoded.