[backend] implement an obsscm service type
An obsscm service is a pseudo service that is not based on
a commit. The service result is not prefixed with _service,
but committed as is.

We currently mis-use the 'title' field in the package meta
to configure the obsscm url, this will be changed to some
extra package element.