Rename @notifications_total to @notifications_count
The total refers to the count of all notifications of a user. It's not
the count of the notifications to be returned to the user.
`@notifications_per_page` isn't valid name here since it refers to how
many notifications are returned per page. This changes depending on the
`show_all` query parameter.