Change the vendoring path
We use workspaces to save transfer artifacts between CircleCI jobs
steps.

On the `install dependencies` step of the `checkout_code` job we
install the gems present on the `Gemfile.lock` system wide, on
`/usr/lib64/ruby2.5.0`. That directory is not being taken into account
when performing the `persist_to_workspace` step, and that's why we get
errors regarding missing gems on the linters.

Changing the vendoring path so it will be included in the workspace
persitence fixes that. When the workspace is restored the gems are there.