Remove unused code in SearchFinder
The `@preloaded_classes` class variable was never modified.

The `@preloads` variable was only assigned for "request" searches. Even
if it considered a typo, and `@preloads` is substituted by
`@preloaded_classes`, this could led to always apply a `preload` to
the final query. This can improve the `GET /search/request` endpoint,
but it is not the desirable behavior for the `GET /search/request/id`
endpoint.