Set optional flag on belongs_to relation where necessary
After enabling `belongs_to_required_by_default` we have to
explicitly tell rails where the associated record is optional.
I set it to optional where ever it is not dissallowed on DB level with the
`not_null` value or where no explicit validation for presence
was set in the past. This way we keep the previous behavior and
don't break any code.