Add data migration to add word boundary to issue tracker regex
Without the word boundary in the regular expression it is possible
that the tracker matches on mentioned issues that contain a certain
sequence of characters that match another tracker as well.
This results in wrongly created issues. We only want to match on
whole words.

Example: "gh#OSInside/kiwi#1650" currently creates two issues in OBS,
"gh#OSInside/kiwi#1650" which correctly matches the github tracker and
by accident "i#1650" which matches the OpenOffice issue tracker.

Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>