Use dom_id for mark_as_read_or_unread_button
This is what ActionView uses to associate records with DOM elements.
Let's use it too!

https://api.rubyonrails.org/classes/ActionView/RecordIdentifier.html#method-i-dom_id

The generated ID is slightly different, `dom_id` uses underscores
instead of dashes. So before, it was `update-notification-123`, now it's
`update_notification_123`. It's not an issue, as we don't refer to this
ID in any code.