Don't initialize `NotificationsFinder` twice
Currently we initialize the `NotificationsFinder` once in the
pundit notification policy scope and then use the returned
notifications to initialize another `NotificationsFinder`
object.
This is confusing and makes it harder to track down unefficient
queries and other issues. We should move the logic that fetches
the notifications for the subscribed user from the finder to the
policy since it is related to authorization.