Instrument Workflow and Workflow::Step
With this, I want to know how many workflows are instantiated, then
executed. This will differ a lot if users rely on filters or not. It
will be interesting to know. It's basic, but it's a first step.

I also want to know which steps are the most popular in workflows. It
allows us to see how the adoption is going for newly introduced steps.

RuboCop TODOs were regenerated since refactoring the Workflow model to
make the class smaller is out of scope for this PR.