Use notification pundit policy in order to authorize and fetch a single notification
Currently we query a whole scope in order to receive and authorize a
single notification. Also having a method in a finder class
to simply query a notification by an Id creates a bit of overhead.
Lets simply use the `find` method to fetch the record and a pundit method
to authorize.