Fix flickering spec in features/webui/projects_spec.rb
According to the CircleCI insights about flickering specs, this spec was
failing in about 3% of the last 100 CI runs. It failed when, for an
obscure reason, the flash wasn't set. I have no idea why, but checking
for the project's title on the project#show page of the newly created
  maintenance incident is sufficient to test the whole workflow.

Link to the CircleCI insights:
https://app.circleci.com/insights/github/openSUSE/open-build-service/workflows/test_all/tests

Fixes #10233