Show link to add a package to the watchlist for ...
... multibuild packages.

Before, an error was thrown when browsing a page where the package was
multibuild. For example, in the live build log.

Fix the retrieval of the package, showing a link to add the root
multibuild package to the watchlist.

Also retrieval of build logs of remote projects was not supported.
Remove related code.

Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>