Don't rely on other conditions to stop a live log
Whether the log chunk size is zero or the request is not the first, the
live log must not stop. Only rely on the list of returning codes
recognized as final states to stop the live log.

This prevents the case of stopping the live log when the log chunk size
is zero, the request is not the first one, and the code returned by the
API is "finished". "finished" is not a final state, and the live log
page should keep requesting AJAX calls until the code returned is a
final one (like "succeeded", "failed", ...).