Respond with an error when the SCM is unsupported
Before, validate_params was only skipped when the SCM is supported
(GitHub or GitLab). This caused issues whenever a payload from an
unsupported SCM didn't pass the checks in validate_params.  For example,
Codeberg seems to include `nil` instead of `null` in its JSON payloads.
This caused validate_params to raise an exception, preventing us from
responding back to the SCM with a helpful error message stating that
Codeberg isn't supported.

Fixes #11861