[backend] add "--filelists_ext" parameter for createrepo
If "filelists_ext" option is enabled for the rpmmd repository, it will
call createrepo_c with the --filelists_ext parameter.

This parameter will create a new optional variant of filelists.xml,
named filelists_ext.xml.  This will contain, besides the list of files,
the hash (usually SHA256) registered in the RPM header for this file
(only for file types, not for ghost or dir types).

Additionally, a new field "checksum" will register the hash algorithm
used by librpm for those file entries.

Signed-off-by: Alberto Planas <aplanas@suse.com>